fix: ECDSA signature DER encoding detection
I was previously assumed the secp251k1 libary would return an error
when trying to parse a non DER encoded signature with `parse_der`.
This is not the case.

Fixed by implementing the BIP66 DER encoding reference:
https://github.com/bitcoin/bips/blob/master/bip-0066.mediawiki#der-encoding-reference
and using it when checking if a ECDSA signature encoded in DER.

Additionally, introduce an DEREncoding enum which can provide more
insights into why a ECDSA signature is not valid DER.

Along with this change, the test vectors are fixed:
1. They incorrectly classified some non-DER encoded signatures as DER
encoded
2. Due to using parse_der() on a non-DER encoded signature, the  r and S
values were wrong, which caused the low_r() and low_s() functions to
return incorrect results.