fix: transaction payment approximation
A "payment" is a non-change output. This is approximated by substracting
1 from the output count if the transaction has more than one output.

However, this wasn't actually implemented. The number of payments was
always the number of outputs. This was an (untested) implementation bug.

This is fixed by actually substracting one from the output count.
Additionally, because the previous number was broken, now opreturn
outputs aren't counted as payments. To make users aware of this, the
API breaks and payments() is now a function on TxInfo instead of a
part of the TxInfo struct.

Also adding test-coverage with a few OP_RETURN edge cases for this.